minor double vs float

+++ Examples/Algorithms/TruthTracking/ActsExamples/TruthTracking/TrackTruthMatcher.cpp
@@ -92,10 +92,10 @@ ActsExamples::ProcessCode TrackTruthMatcher::execute(
     // Check if the trajectory is matched with truth.
     // If not, it will be classified as 'fake'
     const bool recoMatched =
-        static_cast<float>(nMajorityHits) / track.nMeasurements() >=
+        static_cast<double>(nMajorityHits) / track.nMeasurements() >=
         m_cfg.matchingRatio;
     const bool truthMatched =
-        static_cast<float>(nMajorityHits) /
+        static_cast<double>(nMajorityHits) /
             particleTruthHitCount.at(majorityParticleId) >=
         m_cfg.matchingRatio;
